--- app/api/dummyDonationProject.ts
@@ -1,34 +1,0 @@
-// import type { Project } from "@/types/project/project";
-
-// export async function getDonationProjectById(
-//   id: string,
-// ): Promise<Project | null> {
-//   return {
-//     id,
-//     type: "donation",
-//     title: "포비 후원 프로젝트",
-//     subtitle: "우리집 강아지에게 맛있는 간식을",
-//     description: `후원 목적
-// - 오리목뼈 구입
-// - 신라면 블랙 구입
-// - 맥북 구입`,
-//     category: "donation",
-//     images: ["/dog-hero.jpg", "/dog-1.jpg", "/dog-2.jpg"],
-//     creator: {
-//       name: "포비아빠",
-//       avatar: "/shelter-logo.svg",
-//       bio: "행복하게 살아요.",
-//       github: "",
-//       projects: 12,
-//       followers: 1395,
-//     },
-//     updates: [],
-//     reviews: [],
-//     faqs: [],
-//     stats: {
-//       likes: 47,
-//       shares: 12,
-//       views: 1023,
-//     },
-//   };
-// }
